refactor: 퀴즈 컴포넌트 기능별 디렉토리 분리 및 구조 개선
