+++ src/components/quiz/playing/ChoiceQuiz.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../../styles/components/Quiz.css';
+import '../../../styles/components/Quiz.css';
 
 const ChoiceQuiz = ({ currentCharacter, answered, isCorrect, choices, selectedChoice, setSelectedChoice, handleSubmit, handleNext, autoSubmit, setAutoSubmit }) => {
   return (
