+++ src/components/quiz/playing/InputQuiz.jsx
@@ -1,5 +1,5 @@
 import React, { useRef } from 'react';
-import '../../styles/components/Quiz.css';
+import '../../../styles/components/Quiz.css';
 
 const InputQuiz = ({ currentCharacter, answered, isCorrect, userAnswer, setUserAnswer, handleSubmit, handleNext, inputRef }) => {
   return (
